Fix the total points calculation in MatchRrController; Add round-robin match deletion in TournamentController and StageController

+++ tournament-app-server/DTOs/StageDTO.cs
@@ -17,9 +17,9 @@
         public short? third_place_match_number_of_legs { get; set; }
         public short? third_place_match_best_of { get; set; }
         public string? description { get; set; }
-        public double? win_point { get; set; }
-        public double? draw_point { get; set; }
-        public double? lose_point { get; set; }
+        public decimal? win_point { get; set; }
+        public decimal? draw_point { get; set; }
+        public decimal? lose_point { get; set; }
         public string[]? other_criteria_names { get; set; }
         public string[]? other_criteria_sort_direction { get; set; }
     }
